Moved common types to separated file
