+++ src/types/common.ts
@@ -1,0 +1,16 @@
+export type Point = {
+  x: number;
+  y: number;
+};
+
+export type GameSession = {
+  id: string;
+  name: string;
+  difficulty: number;
+  score: number;
+};
+
+export type NewSessionData = {
+  name: string;
+  difficulty: number;
+};
